Fixing packaging issue
