--- hubspaceng/__init__.py
@@ -1,2 +1,0 @@
-"""Define module-level imports."""
-from .api import login  # noqa
