--- hubspaceng/models/devices/__init__.py
@@ -1,6 +1,0 @@
-"""Convenience loader for devices"""
-from .base import BaseDevice
-from .combo import ComboDevice
-from .fan import FanDevice
-from .lock import LockDevice
-from .plug import PlugDevice
