--- hubspaceng/models/devices/lights/__init__.py
@@ -1,4 +1,0 @@
-"""Convenience loader for light devices"""
-from .base import BaseLightDevice
-from .tunable import TunableLightDevice
-from .rgb import RGBLightDevice
