+++ setup.cfg
@@ -4,11 +4,11 @@ license_files=LICENSE
 
 [options]
 package_dir =
-    =hubspaceng
+    =src
 packages = find_namespace:
 
 [options.packages.find]
-where = hubspaceng
+where = src
 
 [options.package_data]
 * =
