+++ src/hubspaceng/__init__.py
@@ -1,0 +1,2 @@
+"""Define module-level imports."""
+from .api import login  # noqa
