+++ src/hubspaceng/models/devices/__init__.py
@@ -1,0 +1,6 @@
+"""Convenience loader for devices"""
+from .base import BaseDevice
+from .combo import ComboDevice
+from .fan import FanDevice
+from .lock import LockDevice
+from .plug import PlugDevice
