+++ src/hubspaceng/models/places/__init__.py
@@ -1,0 +1,4 @@
+"""Convenience loader for places"""
+from .home import Home
+from .place import Place
+from .room import Room
