+++ src/hubspaceng/util.py
@@ -1,0 +1,9 @@
+"""Utility/convenience functions for the Hubspace API"""
+import calendar
+import datetime
+
+def get_utc_time() -> int:
+    """Get current UTC time"""
+    date = datetime.datetime.utcnow()
+    utc_time = calendar.timegm(date.utctimetuple()) * 1000
+    return utc_time
